% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Compute the standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{Numeric, a vector of numbers.}
}
\value{
The standard error of the mean of the numbers in x
}
\description{
Compute the standard error of the mean
}
\examples{
sem(c(1,2,3,4,5))
}
